Fix handling of `--allow-any-qis` param
This channge wires up the internal handling of the `--allow-any-qis` parameter, which was present but not correctly checked before. Now, when the flag is provided, it will override the settings from the loaded adaptor profile and skip checking the qis whitelist, allowing for easier experimentation with new instructions before formally added them to an adaptor profile.