Add disable-straightline-code-requirement option (#161)
This change adds a disable-straightline-code-requirement option. This new option controls whether QIR should be validated to have just one function implementation with a single block.

Before this change, it the behavior was as if this option was set to false by default and only set to true if both the replace-qubit-on-reset and defer-measurements options are set to true.